Breaking change for frontend
Use expo go instead of react native cli. Reason: React native cli is a pain in the azz. React-native-vision-camera usually breaks when it is together with react-native-worklets-core (with a ratio about 7:3, 7 is the break one). Doesn't seem to find a good way to detect face (there's one in the react-native-vision-camera community, but too old and not compatible with vision-camera v3. React-native-fast-tflite only supports ios, the repo's owner already has plan to add the android version but not now, I think. And tfjs doesn't seem to support react native for phone in general (iirc, it only supports web-base). So I move to expo go, use the camera there, with a built-in face detection from google and implement other things like login with face, register with face... Just the base, though. Still need to update backend before finishing it.