Optimize API (or sth like that) + Change the screens' behaviors a bit.
For backend, add note for better use of ngrok (change region). For backend, find a way to replace the use of tunnel (which is unstable) that is localhost. Optimize the API by resize the image before sending (note: The most optimized way should be change the size of the camera, but getting the camera's size takes so much time). Also, change the face register and the login face a bit.